requirements.txt
Flask==3.0.2
keras==3.1.1
numpy==1.26.4
matplotlib==3.8.3
tensorflow==2.16.1
pillow==10.2.0
gunicorn==21.2.0